fix(types): BaoIoRequest return value
This commit is based on replacing the return type of
the BaoIoRequest struct (u64 -> i32) in order to handle
error cases (when the return value is less than zero).

+++ src/api/src/types.rs
@@ -33,7 +33,7 @@ pub struct BaoIoRequest {
     pub access_width: u64,
     pub cpu_id: u64,
     pub vcpu_id: u64,
-    pub ret: u64,
+    pub ret: i32,
 }
 
 /// Struct representing a Bao I/O event file descriptor.
